Remove redundant type specifiers
This removes generic types and casts which aren’t necessary.

JIRA: OPNFLWPLUG-1044

+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/connection/OutboundQueueProviderImplTest.java
@@ -42,7 +42,7 @@ public class OutboundQueueProviderImplTest extends TestCase {
     public void testCreateBarrierRequest() {
         final BarrierInput barrierRequest = outboundQueueProvider.createBarrierRequest(DUMMY_XID);
         assertNotNull(barrierRequest);
-        assertEquals((short)OFConstants.OFP_VERSION_1_3, (short)barrierRequest.getVersion());
+        assertEquals(OFConstants.OFP_VERSION_1_3, (short)barrierRequest.getVersion());
         assertEquals(DUMMY_XID, barrierRequest.getXid());
     }
 }
